Added some basic tests for Problem class
